Updated section on reading data from project.clj
Re-arranged and expanded on the code that can be used to read the data
in the Clojure project.clj file, defining the project.

Broke down the code into individual functions, wrapping one function at
a time around the previous.

Provided a clearer explaination of the threading macro, using the ,,,
notation to show where resutls of one function are passed to the next.
Hopefully showing how code can be made easier to read in this way.