Very simple threading macro example with str
Added very simple and contrived examples of the threading macro using
several calls to the `str` function.

In the first example, each str function joins its own stings together then passes that
combined string to the next function as its first argument.

In the second example we show the difference when using the ->>
threading macro, which passes functuion return values as the last
argument.  This time the string comes out backwards.