Map reduce sandwich example
In the picture at
http://www.datasciencecentral.com/forum/topics/what-is-map-reduce
a sandwich is used as an example to explain map reduce.

Map is used to slide a collection of ingredients.
Reduce is used to combine those ingredients together to form one sandwich.

This is a code example created by @jr0cket to show how the sandwich map reduce
could be implemented in Clojure.