Solution #83 - A half truth
Walk through of the reasoning behind several solutions that were evolved during
a code dojo.

Initially a verbose approach was taken that was fairly simple to follow.

Additional refinements were made that were exactly the same, just with less
code.  First dropping the let expression and then using the short-hand for an
anonymous function.

Finally on the way home I though of a really simple answer, using the not=
function.  Working through the problem was valuable to help fully understand the
algorithm.