4Clojure #26 - Fibonacci Sequence
Analysis of the classic problem of generating the Fibonacci Sequence.  The
approach taken was to add the last two values of the sequence to create the next
number, starting with a seed value of [1 1].

Detailed how the design was evolved using a simple loop recur iteration, showing
the incremental changes in the design until an answer was found.

The local names in the code base were kept of reasonable size to be meaningful,
but could have been converted to single characters to get a slightly lower golf
score (although that isn't important).