4Clojure #61 - Map Construction - decomposing zipmap
I like the 4Clojure problems where you have a restriction, as its a great
opportunity to learn about the restricted function (or functions).

In this example, zipmap is restricted as it is the exact function that would
solve the problem.

By looking at the implementation of zipmap and analysing the problem, a design
was evolved that solved the challenge and helped us understand how zipmap works.
This will be useful in understanding when to apply zipmap.

The incremental steps taken in this design show how conj and reduce functions
can work together as well as how they can be substituted by the into function.