Update: ascii code generator - simplified
Use of a for loop is not required as map can be used to join two collections
together, in this case the collection that is the English language alphabet and
the generated range of ascii characters.

The range of ascii characters is a sequential whole number series, starting from
the given ascii code.  So to generate the ascii number collection, range
function is provided the start code value and a calculation of the end code
value.